# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Agent Planner API",
    "version": "2.0.0",
    "description": "Collaborative planning API for humans and AI agents"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
...
  },
  "paths": {
...
    "/v1/briefing": {
      "get": {
        "summary": "Bundled mission-control state — goal health, pending decisions, active claims, recent activity, top recommendation",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Agent briefing"
          }
        }
      }
    },
    "/v1/tasks/claim-next": {
      "post": {
        "summary": "Pick the next task in scope, claim it, and return its context in one call",
        "description": "Alias of the agent-loop work-session start. Accepts plan_id/goal_id scope, ttl_minutes, depth, dry_run, fresh.",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Work session started (session_id",
            "task": null,
            "claim": null,
            "context)": null
          },
          "409": {
            "description": "Task is already claimed"
          }
        }
      }
    },
    "/v1/tasks/{nodeId}/context": {
      "get": {
        "summary": "Progressive task context (depth 1-4, token-budgeted)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "depth",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 4
            }
          },
          {
            "in": "query",
            "name": "token_budget",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assembled context"
          }
        }
      }
    },
    "/v1/tasks/{nodeId}/update": {
      "post": {
        "summary": "Atomic task state transition — status + log + claim release + learning in one call",
        "description": "Composed facade mirroring the MCP `update_task` tool. Steps apply independently; per-step results are reported in applied/failures. Claim release defaults to automatic when status is completed or blocked.",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Applied steps and per-step failures"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/v1/tasks/{nodeId}/claim": {
      "post": {
        "summary": "Explicitly claim a task (lease-based lock)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Claim created"
          },
          "409": {
            "description": "Task is already claimed"
          }
        }
      },
      "delete": {
        "summary": "Release a task claim",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Claim released"
          }
        }
      }
    },
...
  },
...
}